% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{fingerprint}
\alias{fingerprint}
\alias{fingerprint,PGP-method}
\title{fingerprint}
\usage{
fingerprint(object)

\S4method{fingerprint}{PGP}(object)
}
\arguments{
\item{object}{a \code{PGP} S4 object}
}
\description{
Get the \verb{PGP KEY} fingerprint value
}
