% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{id}
\alias{id}
\alias{id,PGP-method}
\title{id}
\usage{
id(object)

\S4method{id}{PGP}(object)
}
\arguments{
\item{object}{a \code{PGP} S4 object}
}
\description{
Get the \verb{PGP KEY} ID
}
