% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{meta}
\alias{meta}
\alias{meta,PGP-method}
\title{meta}
\usage{
meta(object)

\S4method{meta}{PGP}(object)
}
\arguments{
\item{object}{a `PGP` S4 object}
}
\description{
Get the associated meta data
}
