% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\docType{methods}
\name{pgp}
\alias{pgp}
\alias{pgp,PGP-method}
\title{Accessor Methods for the PGP S4 class}
\usage{
pgp(object)

\S4method{pgp}{PGP}(object)
}
\arguments{
\item{object}{a `PGP` S4 object}
}
\description{
Get the `PGP KEY`
}
