% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_email.R
\name{search_email}
\alias{search_email}
\title{Search the Mailvelope Key Server by Email}
\usage{
search_email(email)
}
\arguments{
\item{email}{a character string of an email address to search on the Mailvelope Key Server}
}
\value{
a `PGP` S4 object
}
\description{
Search the Mailvelope Key Server by Email
}
