% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\docType{methods}
\name{write_key}
\alias{write_key}
\alias{write_key,PGP-method}
\title{write_key}
\usage{
write_key(object, outpath)

\S4method{write_key}{PGP}(object, outpath)
}
\arguments{
\item{object}{a `PGP` S4 object}

\item{outpath}{a valid directory where the key will be saved to}
}
\description{
Write the `PGP_KEY` to a file
}
